+++ expyriment/stimuli/_audio.py
@@ -117,7 +117,7 @@ class Audio(Stimulus):
             self._file = None
             self._is_preloaded = False
 
-    def play(self, loops=0, maxtime=0, fade_ms=0):
+    def play(self, loops=0, maxtime=0, fade_ms=0, log_event_tag=None):
         """Play the audio stimulus.
 
         The function returns immediately after the sound started to play.
@@ -131,6 +131,10 @@ class Audio(Stimulus):
             stop after given amount of milliseconds (default = 0)
         fade_ms : int, optional
             fade in time in milliseconds (default = 0)
+        log_event_tag : numeral or string, optional
+            if log_event_tag is defined and if logging is switched on for this
+            stimulus (default), a summary of the inter-event-intervalls are
+            appended at the end of the event file
 
         """
 
@@ -145,7 +149,8 @@ class Audio(Stimulus):
                 filename = self._filename
 
             _internals.active_exp._event_file_log(
-                "Stimulus,played,{0}".format(filename), 1)
+                "Stimulus,played,{0}".format(filename), 1,
+                                 log_event_tag=log_event_tag)
         return rtn
 
     def stop(self):
@@ -154,7 +159,7 @@ class Audio(Stimulus):
         if self._is_preloaded:
             self._file.stop()
 
-    def present(self):
+    def present(self, log_event_tag=None):
         """Presents the sound.
 
         The function is identical to Audio.play(loops=0, maxtime=0, fade_ms=0)
@@ -166,4 +171,4 @@ class Audio(Stimulus):
 
         """
 
-        self.play()
+        self.play(log_event_tag=log_event_tag)